BosonSampling.compute_confidence — Method

compute_confidence(events,p_q, p_a)

A bayesian confidence estimator: return the probability that the null hypothesis Q is right compared to the alternative hypothesis A.

BosonSampling.compute_confidence_array — Method

compute_confidence_array(events, p_q, p_a)

Return an array of the probabilities of H being true as we process more and more events.

BosonSampling.compute_probability! — Method

compute_probability!(b::Bayesian)

Updates all probabilities associated with a Bayesian Certifier.

BosonSampling.number_of_samples — Method

number_of_samples(evb::Event{TIn, TOut}, evd::Event{TIn, TOut}; p_null = 0.95, maxiter = 10000) where {TIn <:InputType, TOut <:PartitionCountsAll}

Outputs the number of samples required to attain a confidence that the null hypothesis (underlied by the parameters sent in evb) is true compared the alternative (underlied by evd) through a bayesian partition sample.

Note that this gives a specific sample - this function should be averaged over many trials to obtain a reliable estimate.

BosonSampling.p_B — Method

p_B(event::Event{TIn, TOut}) where {TIn<:InputType, TOut <: FockDetection}

Outputs the probability that a given FockDetection would have if the InputType was Bosonic for this event.

BosonSampling.p_D — Method

p_D(event::Event{TIn, TOut}) where {TIn<:InputType, TOut <: FockDetection}

Outputs the probability that a given FockDetection would have if the InputType was Distinguishable for this event.