BosonSampling.noisy_distribution — Method

noisy_distribution(;input::Input, loss::Real, interf::Interferometer, exact=true, approx=true, samp=true, error=1e-4, failure_probability=1e-4)

Compute the exact and/or approximated and/or sampled probability distribution of all possible output configurations of partially-distinguishable photons through a lossy interferometer. By default, exact, approx and samp are set to true meaning that noisy_distribution returns an array containing the three distributions.

Note

The probabilities within a distribution are indexed following the same order as output_mode_occupation(n,m)

The approximated distribution has error and failure probability of $1e^{-4}$.

Reference

https://arxiv.org/pdf/1809.01953.pdf

BosonSampling.theoretical_distribution — Method

theoretical_distribution(;input::Input, interf::Interferometer, i=nothing)

Compute the probability distribution of all possible output configurations of fully/partially-indistinguishable photons through a lossless interferometer.

Note

The probabilities within the distribution are indexed following the same order as output_mode_occupation(n,m)

If i (with default value nothing) is set to an integer theoretical_distribution returns the probability to find the photons in the i'th configuration given by output_mode_occupation