BosonSampling.clean_pdf — Function

clean_pdf(A::Array, atol=ATOL)

Checks wether an array is an acceptable discrete probability distribution with tolerance ATOL. If so, converts its elements to normalized positive real numbers.

BosonSampling.clean_proba — Function

clean_proba_(probability:Number, atol=ATOL)

Checks whether a (complex) number is close enough to a valid probability with tolerance ATOL. If so, convert it to a positive real number.

BosonSampling.isa_pdf — Method

isa_pdf(pdf)

Asserts if pdf is a valid probability distribution.

BosonSampling.sqr — Method

sqr(a,b)

Computes the euclidian distance between two probability distributions.

BosonSampling.tvd — Method

tvd(a,b)

Computes the total variation distance between two probability distributions.