BosonSampling.fourier_matrix — Method

fourier_matrix(n::Int; normalized=true)

Returns a $n$-by-$n$ Fourier matrix with optional normalization (true by default).

BosonSampling.gram_matrix_one_param — Method

gram_matrix_one_param(n::Int, x::Real)

Returns a $n$-by-$n$ Gram matrix parametrized by the real $0 ≤ x ≦ 1$.

BosonSampling.hadamard_matrix — Function

hadamard_matrix(n::Int; normalized=true)

Returns a $n$-by-$n$ Hadamard matrix with optional normalization (true by default).

BosonSampling.perturbed_gram_matrix — Method

perturbed_gram_matrix(M, epsilon)

Returns a Gram matrix generated by the columns of M which are perturbed by a Gaussian quantity of variance epsilon once normalized.

BosonSampling.perturbed_unitary — Method

perturbed_unitary(U, epsilon)

Returns a unitary matrix whose columns are generating vector perturbed by a random Gaussian quantity with variance epsilon once normalized.

BosonSampling.rand_gram_matrix — Method

rand_gram_matrix(n::Int)

Returns a full rank $n$-by-$n$ random Gram matrix.

BosonSampling.rand_gram_matrix_from_orthonormal_basis — Method

rand_gram_matrix_from_orthonormal_basis(n::Int, r::Int)

Returns a $n$-by-$n$ random Gram matrix that generates orthonormal basis of r vectors.

BosonSampling.rand_gram_matrix_positive — Method

rand_gram_matrix_positive(n::Int)

Returns a positive elements $n$-by-$n$ random Gram matrix of full rank.

BosonSampling.rand_gram_matrix_rank — Method

rand_gram_matrix_rank(n::Int, r::Int)

Returns a $n$-by-$n$ random Gram matrix of maximum rank and great likelihood r.

BosonSampling.rand_gram_matrix_real — Method

rand_gram_matrix_real(n::Int)

Returns a real $n$-by-$n$ random Gram matrix of full rank.

BosonSampling.rand_haar — Method

rand_haar(n::Int)

Returns a $n$-by-$n$ Haar distributed unitary matrix following https://case.edu/artsci/math/esmeckes/MeckesSAMSILecture2.pdf.

BosonSampling.sylvester_matrix — Method

sylvester_matrix(p::Int; normalized=true)

Returns a $2^p$-by-$2^p$ Sylvester matrix with optional normalization (true by default) following https://arxiv.org/abs/1502.06372.