BosonSampling.Event — Type

Event{TIn<:InputType, TOut<:OutputMeasurementType}

Event linking an input to an output.

Fields:
	- input_state::Input{TIn}
	- output_measurement::TOut
	- proba_params::EventProbability
	- interferometer::Interferometer

BosonSampling.EventProbability — Type

EventProbability(probability::Union{Nothing, Number})
EventProbability(mc::MultipleCounts)

Holds the probability or probabilities of an Event.

Fields:
	- probability::Union{Number,Nothing, MultipleCounts}
	- precision::Union{Number,Nothing}
	- failure_probability::Union{Number,Nothing}

BosonSampling.MultipleCounts — Type

MultipleCounts()
MultipleCounts(counts, proba)

Holds something like the photon counting probabilities with their respective probability (in order to use them as a single observation). Can be declared empty as a placeholder.

Fields:
	- counts::Union{Nothing, Vector{ModeOccupation}, Vector{PartitionOccupancy}},
	- proba::Union{Nothing,Vector{Real}}