BosonSampling.Bosonic — Type

Type used to notify that the input is made of FockState indistiguishable photons.

BosonSampling.CoherentState — Type

Type used to notify that the input is made of coherent state.

BosonSampling.Distinguishable — Type

Model of distinguishable photons FockState.

BosonSampling.Gaussian — Type

Type used to notify that the input is made of Gaussian states.

BosonSampling.GramMatrix — Type

GramMatrix{T}(n::Int) where {T<:InputType}
GramMatrix{T}(n::Int, distinguishability_param::Real) where {T<:InputType}
GramMatrix{T}(n::Int, S::Matrix) where {T<:InputType}

Matrix of partial distinguishability. Will automatically generate the proper matrix related to the provided InputType.

Fields:
    - n::Int: photons number
    - S::Matrix: Gram matrix
    - rank::Union{Int, Nothing}
    - distinguishability_param::Union{Real, Nothing}
    - generating_vectors::OrthonormalBasis

BosonSampling.Input — Type

Input{T<:InputType}
Input{T}(r::ModeOccupation) where {T<:InputType}
Input{T}(r::ModeOccupation, G::GramMatrix) where {T<:InputType}

Input state at the entrance of the interferometer.

Fields:
    - r::ModeOccupation
    - n::Int
    - m::Int: modes numbers
    - G::GramMatrix
    - distinguishability_param::Union{Real, Nothing}

BosonSampling.InputType — Type

Supertype to any concrete input type such as Bosonic, PartDist, Distinguishable and Undef.

BosonSampling.OneParameterInterpolation — Type

One parameter model of partial distinguishability interpolating between indistinguishable and fully distinguishable photons FockState.

Reference

Sampling of partially distinguishable bosons and the relation to the multidimensional permanent

BosonSampling.OrthonormalBasis — Type

OrthonormalBasis(vector_matrix::Union{Matrix, Nothing})

Basis of vectors $v_1,...,v_n$ stored as columns in a $n$-by-$r$ matrix possibly empty.

Fields:
    vectors_matrix::Union{Matrix,Nothing}

BosonSampling.PartDist — Type

Type used to notify that the input is made of FockState partially distinguishable photons.

BosonSampling.RandomGramMatrix — Type

Model of partially distinguishable photons FockState described by a randomly generated GramMatrix.

BosonSampling.SingleModeSqueezedVacuum — Type

Type used to notify that the input is made of single mode squeezed state.

BosonSampling.ThermalState — Type

Type used to notify that the input is made of thermal state.

BosonSampling.Undef — Type

Model of photons FockState with undefined GramMatrix.

BosonSampling.UserDefinedGramMatrix — Type

Model of partially distinguishable photons FockState described by a provided GramMatrix.

BosonSampling.VacuumState — Type

Type used to notify that the input is made of the vacuum state.