BosonSampling.Fourier — Type

Fourier(m::Int)

Creates a Fourier Interferometer of dimension m.

Fields:
    - m::Int
    - U::Matrix

BosonSampling.Hadamard — Type

Hadamard(m::Int)

Creates a Hadamard Interferometer of dimension m.

Fields:
    - m::Int
    - U::Matrix

BosonSampling.Interferometer — Type

Supertype to any concrete interferomter type such as UserDefinedInterferometer, RandHaar, Fourier,...

BosonSampling.LossParameters — Type

IsLossy{T}
IsLossless{T}

Trait to refer to quantities with inclusion of loss: the real Interferometer has dimension m_real * m_real while we model it by a 2m_real * 2m_real one where the last m_real modes are environment modes containing the lost photons.

BosonSampling.RandHaar — Type

RandHaar(m::Int)

Creates an instance of Interferometer from a Haar distributed unitary matrix of dimension m.

Fields:
    - m::Int
    - U::Matrix

BosonSampling.UserDefinedInterferometer — Type

UserDefinedInterferometer(U::Matrix)

Creates an instance of Interferometer from a given unitary matrix U.

Fields:
    - m::Int
    - U::Matrix