BosonSampling.get_parametric_type — Method

get_parametric_type(i)

Return the types T1,..,Tn of a parametric type i::T{T1,..,Tn}.

Note

If the parametric type has only one parameter, use get_parametric_type(i)[1].

If no parametric type, returns an array containing the type itself.