BosonSampling.classical_sampler — Method

classical_sampler(U, n, m)
classical_sampler(;input::Input, interf::Interferometer)

Sample photons according to the Distinguishable case.

BosonSampling.cliffords_sampler — Method

cliffords_sampler(;input::Input, interf::Interferometer)

Sample photons according to the Bosonic case following Clifford & Clifford algorithm performed (at most) in $O(n2^m + Poly(n,m))$ time and $O(m)$ space.

BosonSampling.metropolis_sampler — Method

metropolis_sampler(;target_pdf, known_pdf, known_sampler, starting_state, n_iter, n_burn = 100, n_thinning = 100)

Implement a metropolis independent sampler for standard boson sampling following. The burn in period n_burnand the thinning interval n_thinning both have default value of 100.

Reference

https://arxiv.org/abs/1705.00686: As the paper is limited to collinionless events, we keep track of this thanks to iterate_until_collisionless.

Warning

Burn in perdiod and thinning interval must have the same value.

BosonSampling.noisy_sampler — Method

noisy_sampler(;input::Input, reflectivity::Real, interf::Interferometer)

Sample partially-distinguishable photons through a lossy interferometer, which runs (at most) in $O(n2^m + Poly(n,m))$ time.

Reference

https://arxiv.org/pdf/1907.00022.pdf