BosonSampling.BeamSplitter — Type

BeamSplitter(transmission_amplitude::Float64)

Creates a beam-splitter with tunable transmissivity.

Fields:
    - transmission_amplitude::Float64
    - U::Matrix{ComplexF64}
    - m::Int

BosonSampling.Fourier — Type

Fourier(m::Int)

Creates a Fourier Interferometer of dimension m.

Fields:
    - m::Int
    - U::Matrix

BosonSampling.Hadamard — Type

Hadamard(m::Int)

Creates a Hadamard Interferometer of dimension m.

Fields:
    - m::Int
    - U::Matrix

BosonSampling.Interferometer — Type

Supertype to any concrete interferomter type such as UserDefinedInterferometer, RandHaar, Fourier,...

BosonSampling.PhaseShift — Type

PhaseShift(shifted_modes::Array, param_::Array)

Creates a phase-shifter that is applied on the modes precised by shiftedmodes with phase shifts given in param.

Fields:
    - shifted_modes::Array
    - param_::Array
    - m::Int
    - U::Matrix{ComplexF64}

BosonSampling.RandHaar — Type

RandHaar(m::Int)

Creates an instance of Interferometer from a Haar distributed unitary matrix of dimension m.

Fields:
    - m::Int
    - U::Matrix

BosonSampling.Rotation — Type

Rotation(angle::Float64)

Creates a Rotation matrix with tunable angle.

Fields:
    - angle::Float64
    - U::Matrix{ComplexF64}
    - m::Int

BosonSampling.UserDefinedInterferometer — Type

UserDefinedInterferometer(U::Matrix)

Creates an instance of Interferometer from a given unitary matrix U.

Fields:
    - m::Int
    - U::Matrix